break-it-down-ai/pyproject.toml
[project]
name = "break-it-down-ai"
version = "0.1.0"
description = "FastAPI service for breaking tasks into steps"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn==0.30.6",
  "pydantic==2.9.2",
  "openai==1.54.4",
  "python-dotenv==1.0.1",
  "httpx==0.27.2",
]

[tool.uvicorn]
factory = false
